Do not use player UIDs for persistence (#9)
I previously assumed that a player would always have
the same UID, but you only have to dis- and reconnect
once to see that that's not true.

This also uncovered a bug switching up sector &
cp times.

+++ gbx/src/adapter/callbacks.rs
@@ -73,7 +73,7 @@ fn forward_regular_callback(cb_out: &Sender<Callback>, call: Call) -> CallbackTy
             if let [Int(uid), String(login), String(msg), Bool(_is_registered_cmd)] = &call.args[..]
             {
                 return success(PlayerChat {
-                    from_uid: *uid as i32,
+                    from_uid: *uid,
                     from_login: login.clone(),
                     message: msg.clone(),
                 });
@@ -96,7 +96,7 @@ fn forward_regular_callback(cb_out: &Sender<Callback>, call: Call) -> CallbackTy
         "ManiaPlanet.PlayerManialinkPageAnswer" => {
             if let [Int(uid), String(login), String(answer), Array(_entries)] = &call.args[..] {
                 return success(PlayerAnswer {
-                    from_uid: *uid as i32,
+                    from_uid: *uid,
                     from_login: login.clone(),
                     answer: answer.clone(),
                 });
